+++ src/de/hhu/alobe/ba2016/editor/aktionen/Aktion_NeuesLinsenFormat.java
@@ -7,18 +7,18 @@ public class Aktion_NeuesLinsenFormat extends Aktion {
     Linse linse;
 
     private double alt_brechzahl;
-    private float alt_hoehe;
-    private float alt_dicke;
-    private float alt_radius1;
-    private float alt_radius2;
+    private double alt_hoehe;
+    private double alt_dicke;
+    private double alt_radius1;
+    private double alt_radius2;
 
     private double neu_brechzahl;
-    private float neu_hoehe;
-    private float neu_dicke;
-    private float neu_radius1;
-    private float neu_radius2;
+    private double neu_hoehe;
+    private double neu_dicke;
+    private double neu_radius1;
+    private double neu_radius2;
     
-    public Aktion_NeuesLinsenFormat (Linse linse, double brechzahl, float hoehe, float dicke, float radius1, float radius2) {
+    public Aktion_NeuesLinsenFormat (Linse linse, double brechzahl, double hoehe, double dicke, double radius1, double radius2) {
         this.linse = linse;
 
         alt_brechzahl = linse.getBrechzahl();
