+++ src/de/hhu/alobe/ba2016/physik/elemente/Auge/Netzhaut.java
@@ -6,7 +6,7 @@ import de.hhu.alobe.ba2016.physik.elemente.absorbtion.Schirm;
 
 public class Netzhaut extends Schirm{
 
-    public Netzhaut(Auge auge, Vektor position, int hoehe) {
+    public Netzhaut(Auge auge, Vektor position, double hoehe) {
         super(auge.getOptischeBank(), position, 0, hoehe);
     }
